style(compiler): make repo clippy-clean and gate CI on it (#96)
Resolved all 50 pre-existing clippy warnings (cargo clippy --all-targets
-- -D warnings now exits 0):

- collapsible_if, single_char_add_str, unnecessary_map_or, len_zero,
  clone_on_copy, needless_question_mark, useless_conversion,
  nonminimal_bool, get_first, manual_map_or, needless_range_loop:
  applied via 'cargo clippy --fix' + a few manual touches.
- manual_strip (x4): use str::strip_prefix instead of starts_with+slice.
- ptr_arg: get_field_type(&String) -> get_field_type(&str).
- should_implement_trait: rename Type::from_str -> Type::parse (avoids
  shadowing std::str::FromStr::from_str).
- module_inception: fold src/lexer/lexer.rs into src/lexer/mod.rs.
- if_same_then_else: drop redundant else-if arm in string interpolation.
- assertions_on_constants: assert!(false, ...) -> panic!(...).

Added two CI jobs to .github/workflows/ci.yml:
- clippy: runs 'cargo clippy --all-targets -- -D warnings'
- fmt: runs 'cargo fmt --check'

Both reuse the aion-compiler Docker image. No behavior change — 87
integration + 9 lib tests still pass.

Closes #96.
